API: move mot_and_sig to conftest
